Linux/Mac: Refactor *nix Utilities classes
