+++ volatility/framework/plugins/mac/check_syscall.py
@@ -6,10 +6,10 @@ from typing import List
 
 from volatility.framework import exceptions, interfaces
 from volatility.framework import renderers, contexts
-from volatility.framework.automagic import mac
 from volatility.framework.configuration import requirements
 from volatility.framework.interfaces import plugins
 from volatility.framework.renderers import format_hints
+from volatility.framework.symbols import mac
 from volatility.plugins.mac import lsmod
 
 vollog = logging.getLogger(__name__)
@@ -46,7 +46,7 @@ class Check_syscall(plugins.PluginInterface):
         for (i, ent) in enumerate(table):
             try:
                 call_addr = ent.sy_call.dereference().vol.offset
-            except exceptions.InvalidPagedAddressException:
+            except exceptions.InvalidAddressException:
                 continue
 
             if not call_addr or call_addr == 0:
