+++ volatility/framework/plugins/mac/trustedbsd.py
@@ -7,11 +7,11 @@ from typing import List, Iterator, Any
 
 from volatility.framework import exceptions, interfaces
 from volatility.framework import renderers, contexts
-from volatility.framework.automagic import mac
 from volatility.framework.configuration import requirements
 from volatility.framework.interfaces import plugins
 from volatility.framework.objects import utility
 from volatility.framework.renderers import format_hints
+from volatility.framework.symbols import mac
 from volatility.plugins.mac import lsmod
 
 vollog = logging.getLogger(__name__)
