+++ volatility/framework/symbols/linux/extensions/__init__.py
@@ -8,9 +8,8 @@ from typing import Generator, Iterable, Iterator, Optional, Tuple
 
 from volatility.framework import constants
 from volatility.framework import exceptions, objects, interfaces
-from volatility.framework.automagic import linux
 from volatility.framework.layers import linear
-from volatility.framework.symbols import generic
+from volatility.framework.symbols import generic, linux
 
 vollog = logging.getLogger(__name__)
